requirements.txt
numpy==1.19.2
Pillow==9.1.0
streamlit==1.8.1
tensorflow==2.6.0
keras==2.6.0
opencv-python-headless